Move permission checks into the command runner
