+++ src/commands/queue/restart.test.ts
@@ -12,9 +12,6 @@ const mockUseQueue = useQueue as jest.Mock;
 import getQueueChannel from "../../actions/queue/getQueueChannel";
 const mockGetQueueChannel = getQueueChannel as jest.Mock;
 
-import { userIsAdminForQueueInGuild } from "../../permissions";
-const mockUserIsAdminForQueueInGuild = userIsAdminForQueueInGuild as jest.Mock;
-
 const mockGetAllEntries = jest.fn();
 const mockQueueClear = jest.fn();
 
@@ -25,7 +22,6 @@ import { useTestLogger } from "../../../tests/testUtils/logger";
 
 const mockPrepareForLongRunningTasks = jest.fn().mockResolvedValue(undefined);
 const mockReply = jest.fn().mockResolvedValue(undefined);
-const mockReplyPrivately = jest.fn().mockResolvedValue(undefined);
 
 const logger = useTestLogger("error");
 
@@ -36,12 +32,8 @@ describe("Clear queue contents", () => {
     context = ({
       logger,
       guild: "the guild",
-      channel: {
-        id: "not-queue-channel"
-      },
       prepareForLongRunningTasks: mockPrepareForLongRunningTasks,
-      reply: mockReply,
-      replyPrivately: mockReplyPrivately
+      reply: mockReply
     } as unknown) as GuildedCommandContext;
 
     mockUseQueue.mockReturnValue({
@@ -50,59 +42,26 @@ describe("Clear queue contents", () => {
     });
     mockGetAllEntries.mockResolvedValue([]);
     mockQueueClear.mockResolvedValue(undefined);
-    mockReplyPrivately.mockResolvedValue(undefined);
-    mockUserIsAdminForQueueInGuild.mockResolvedValue(false);
   });
 
-  test("does nothing when admin and no queue is set up", async () => {
+  test("does nothing when no queue is set up", async () => {
     mockGetQueueChannel.mockResolvedValue(null);
-    mockUserIsAdminForQueueInGuild.mockResolvedValue(true);
 
     await expect(restart.execute(context)).resolves.toBeUndefined();
 
-    expect(mockReplyPrivately).not.toHaveBeenCalled();
     expect(mockReply).toHaveBeenCalledTimes(1);
     expect(mockReply).toHaveBeenCalledWith("No queue is set up. Maybe that's what you wanted...?");
     expect(mockUseQueue).not.toHaveBeenCalled();
   });
 
-  test("does nothing when not admin and no queue is set up", async () => {
-    mockGetQueueChannel.mockResolvedValue(null);
-    mockUserIsAdminForQueueInGuild.mockResolvedValue(false);
-
-    await expect(restart.execute(context)).resolves.toBeUndefined();
-
-    // Non-admins outside the queue channel get a permission error
-    expect(mockReply).not.toHaveBeenCalled();
-    expect(mockReplyPrivately).toHaveBeenCalledTimes(1);
-    expect(mockReplyPrivately).toHaveBeenCalledWith(
-      "YOU SHALL NOT PAAAAAASS!\nOr, y'know, something like that..."
-    );
-    expect(mockUseQueue).not.toHaveBeenCalled();
-  });
-
-  test("does nothing when not admin and not in queue channel", async () => {
-    mockGetQueueChannel.mockResolvedValue({ id: "queue-channel" });
-    mockUserIsAdminForQueueInGuild.mockResolvedValue(false);
-
-    await expect(restart.execute(context)).resolves.toBeUndefined();
-
-    expect(mockReply).not.toHaveBeenCalled();
-    expect(mockReplyPrivately).toHaveBeenCalledTimes(1);
-    expect(mockReplyPrivately).toHaveBeenCalledWith(
-      "YOU SHALL NOT PAAAAAASS!\nOr, y'know, something like that..."
-    );
-    expect(mockUseQueue).not.toHaveBeenCalled();
-  });
-
   test.each`
-    channelId              | isAdmin
-    ${"not-queue-channel"} | ${true}
-    ${"queue-channel"}     | ${true}
-    ${"queue-channel"}     | ${false}
+    channelId
+    ${"not-queue-channel"}
+    ${"queue-channel"}
+    ${"queue-channel"}
   `(
     "clears the queue when admin==$isAdmin and in $channelId",
-    async ({ channelId, isAdmin }: { channelId: string; isAdmin: boolean }) => {
+    async ({ channelId }: { channelId: string }) => {
       const queueChannel = { id: "queue-channel" };
       const queueEntries = [
         { queueMessageId: "message1" },
@@ -110,7 +69,6 @@ describe("Clear queue contents", () => {
         { queueMessageId: "message3" }
       ];
       mockGetQueueChannel.mockResolvedValue(queueChannel);
-      mockUserIsAdminForQueueInGuild.mockResolvedValue(isAdmin);
       mockGetAllEntries.mockResolvedValue(queueEntries);
       context.channel = ({
         id: channelId
