+++ src/commands/queue/teardown.test.ts
@@ -1,8 +1,4 @@
-jest.mock("../../permissions");
 jest.mock("../../useGuildStorage");
-
-import { userIsAdminInGuild } from "../../permissions";
-const mockUserIsAdminInGuild = userIsAdminInGuild as jest.Mock;
 
 import { useGuildStorage } from "../../useGuildStorage";
 const mockUseGuildStorage = useGuildStorage as jest.Mock;
@@ -13,7 +9,6 @@ import teardown from "./teardown";
 
 const mockSetQueueChannel = jest.fn();
 const mockReply = jest.fn().mockResolvedValue(undefined);
-const mockReplyPrivately = jest.fn().mockResolvedValue(undefined);
 
 const logger = useTestLogger("error");
 
@@ -24,34 +19,18 @@ describe("Queue teardown", () => {
     context = ({
       guild: "the guild",
       logger,
-      reply: mockReply,
-      replyPrivately: mockReplyPrivately
+      reply: mockReply
     } as unknown) as GuildedCommandContext;
 
-    mockReplyPrivately.mockResolvedValue(undefined);
-    mockUserIsAdminInGuild.mockResolvedValue(true);
     mockUseGuildStorage.mockReturnValue({
       setQueueChannel: mockSetQueueChannel
     });
     mockSetQueueChannel.mockResolvedValue(undefined);
   });
 
-  test("does nothing when the sender doesn't have permission", async () => {
-    mockUserIsAdminInGuild.mockResolvedValue(false);
-    await expect(teardown.execute(context)).resolves.toBeUndefined();
-
-    expect(mockReply).not.toHaveBeenCalled();
-    expect(mockReplyPrivately).toHaveBeenCalledTimes(1);
-    expect(mockReplyPrivately).toHaveBeenCalledWith(
-      "YOU SHALL NOT PAAAAAASS!\nOr, y'know, something like that..."
-    );
-    expect(mockSetQueueChannel).not.toHaveBeenCalled();
-  });
-
   test("unsets the guild queue", async () => {
     await expect(teardown.execute(context)).resolves.toBeUndefined();
 
-    expect(mockReplyPrivately).not.toHaveBeenCalled();
     expect(mockReply).toHaveBeenCalledTimes(1);
     expect(mockReply).toHaveBeenCalledWith("Queue deleted.");
     expect(mockSetQueueChannel).toHaveBeenCalledTimes(1);
