+++ src/handleCommand.ts
@@ -5,7 +5,8 @@ import type Discord from "discord.js";
 import { getEnv } from "./helpers/environment";
 import { getConfigCommandPrefix } from "./actions/config/getConfigValue";
 import { randomGreeting, randomPhrase, randomQuestion } from "./helpers/randomStrings";
-import { allCommands as commands, invokeCommand } from "./commands";
+import { invokeCommand } from "./actions/invokeCommand";
+import { allCommands as commands } from "./commands";
 import { deleteMessage, reply, replyPrivately } from "./actions/messages";
 import getUserIdFromMention from "./helpers/getUserIdFromMention";
 import logUser from "./helpers/logUser";
