+++ src/handleInteraction.ts
@@ -2,7 +2,8 @@ import type Discord from "discord.js";
 import type { CommandContext } from "./commands";
 import type { Logger } from "./logger";
 import type { Storage } from "./configStorage";
-import { allCommands, invokeCommand } from "./commands";
+import { invokeCommand } from "./actions/invokeCommand";
+import { allCommands } from "./commands";
 import { getEnv } from "./helpers/environment";
 import { replyPrivately } from "./actions/messages";
 import logUser from "./helpers/logUser";
